requirements.txt
Flask
pip
numpy
opencv-python
webcolors
pillow